HBX-1682: Move class 'org.hibernate.tool.hbm2x.visitor.HBMTagForPersistentClassVisitor' to 'org.hibernate.tool.internal.export.hbm.HBMTagForPersistentClassVisitor'
